<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/criterion.R
\name{adj.r.squared}
\alias{adj.r.squared}
\title{An rFSA Criterion Function.}
\usage{
adj.r.squared(model, name = "Adj R Squared")
}
\arguments{
\item{model}{lm or glm fit to be passed.}

\item{name}{passed to print.FSA}
}
\description{
rFSA Criterion Function to compute Adjusted R-Squared.
}

=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{adj.r.squared}
\alias{adj.r.squared}
\title{An rFSA Criterion Function.}
\usage{
adj.r.squared(model, name = "Adj R Squared")
}
\arguments{
\item{model}{lm or glm fit to be passed.}

\item{name}{passed to print.FSA}
}
\description{
rFSA Criterion Function to compute Adjusted R-Squared.
}
>>>>>>> 6e2c25ebe54e059b1e387c860c1d0fa8f6e75136
