% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_arc_plc_payments.R
\name{aggregate_multi_param_results}
\alias{aggregate_multi_param_results}
\title{Helper function to aggregate results for multiple parameter combinations}
\usage{
aggregate_multi_param_results(results, aggregate_level, sequestration_rate)
}
\arguments{
\item{results}{Combined results data frame}

\item{aggregate_level}{Aggregation level}

\item{sequestration_rate}{Sequestration rate}
}
\description{
Helper function to aggregate results for multiple parameter combinations
}
\keyword{internal}
