% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{apress}
\alias{apress}
\title{An rFSA Criterion Function.}
\usage{
apress(model, name = "PRESS")
}
\arguments{
\item{model}{lm or glm fit to be passed.}

\item{name}{passed to print.FSA}
}
\description{
rFSA Criterion Function to Allen's Press Statistic.
}
