% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fsaArcIcPrice}
\alias{fsaArcIcPrice}
\title{Commodity-specific ARC-IC benchmark prices, MYA prices, and statutory reference prices.}
\format{
A tibble
\describe{
\item{crop}{Crop name or commodity covered by ARC/PLC.}
\item{marketing_year_dates}{Date range defining the marketing year for the crop (e.g., "Sep. 1–Aug. 31").}
\item{publishing_dates_for_final_mya_price}{Date the final MYA and ARC-IC actual prices were published.}
\item{unit}{Unit of measure for the commodity (e.g., Bushel, Pound).}
\item{reference_price_combined}{Statutory reference price up untill effective reference prices were introduced (following 2018 farm bill), at which point effective reference prices are reported}
\item{annual_benchmark_price_lag5}{Price from five years prior to the marketing year (T-5), used in ARC-CO benchmark calculation.}
\item{annual_benchmark_price_lag4}{Price from four years prior (T-4), used in ARC-CO benchmark calculation.}
\item{annual_benchmark_price_lag3}{Price from three years prior (T-3), used in ARC-CO benchmark calculation.}
\item{annual_benchmark_price_lag2}{Price from two years prior (T-2), used in ARC-CO benchmark calculation.}
\item{annual_benchmark_price_lag1}{Price from one year prior (T-1), used in ARC-CO benchmark calculation.}
\item{current_mya_price}{Marketing Year Average price for the current year (T-0).}
\item{current_national_loan_rate}{National loan rate for the current marketing year.}
\item{current_arcic_actual_price}{Actual ARC-IC price used in revenue calculations for T-0.}
\item{marketing_year}{Marketing year (e.g., "2020–2021").}
\item{program_year}{Program year corresponding to the marketing year (e.g., 2020).}
\item{crop_type}{Subtype of the crop, if applicable (e.g., "long grain", "small").}
\item{rma_type_code}{RMA type code used to classify crop type (may be NA).}
\item{rma_crop_code}{RMA crop code used for identification of commodity.}
}
}
\source{
\url{https://www.fsa.usda.gov/resources/programs/arc-plc/program-data}
}
\usage{
data(fsaArcIcPrice)
}
\description{
Commodity-specific ARC-IC benchmark prices, MYA prices, and statutory reference prices.
}
\details{
To view code used to generate this data set, see \code{./data-raw/fsaArcPlc/supplementary_files/fsaArcIcPrice.R}
}
\keyword{datasets}
