% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fsaCountyBaseAcres}
\alias{fsaCountyBaseAcres}
\title{County‐level base acres and average PLC yields by crop}
\format{
data frame with 99,308 rows and 12 variables:
\describe{
\item{state}{character. Full name of the state.}
\item{county}{character. Name of the county.}
\item{state_code}{character. Two-digit state FIPS code.}
\item{county_code}{character. Three-digit county FIPS code.}
\item{crop}{character. Name of covered commodity}
\item{crop_type}{character. Name of the crop type (e.g. "long grain", "temperate japonica").}
\item{base_acres}{numeric. Base acreage for this crop in the county (acres).}
\item{avg_plc_yield}{numeric. Average PLC yield for the crop.}
\item{program_year}{integer. Program year (e.g. 2014-2023).}
\item{rma_type_code}{character. RMA crop type code.}
\item{rma_crop_code}{integer. RMA crop code.}
\item{fips}{character. Five-digit combined state+county FIPS code.}
}
}
\source{
\url{https://www.fsa.usda.gov/programs-and-services/arcplc_program/index}
}
\usage{
data(fsaCountyBaseAcres)
}
\description{
A dataset containing county‐level base acreage and average PLC yield for each covered
commodity, along with associated RMA and FIPS codes, by program year.
}
\keyword{datasets}
