% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fsaPlcPaymentRate}
\alias{fsaPlcPaymentRate}
\title{Commodity specific PLC payment rates including statutory and effective reference prices.}
\format{
A tibble
\describe{
\item{Commodity}{The name of the commodity}
\item{Marketing Year}{The dates for the commodity specific marketing year }
\item{year}{The two calendar years that the commodity specific marketing year spans}
\item{Publishing Dates for the  Final T - 0 PLC Effective Price}{The date the final T-0 PLC effective price was published}
\item{Unit}{The unit of measurement for the price}
\item{Statutory Reference Price}{The statutory reference price for the commodity}
\item{Final T-0 Effective Price}{The effective reference price for the commodity in the marketing year defined by \code{year}}
\item{Final T-0 MYA Price}{The marketing year average price for the marketing year defined by \code{year}}
\item{T-0 National Loan Rate}{The national loan rate for the commodity in the marketing year defined by \code{year}}
\item{Final T-0 PLC Payment Rate}{The PLC payment rate for the commodity in the marketing year defined by \code{year}}
\item{Maximum PLC Payment Rate}{The maximum PLC payment rate for the commodity in the marketing year defined by \code{year}}
}
}
\source{
\url{https://www.fsa.usda.gov/resources/programs/arc-plc/program-data}
}
\usage{
data(fsaPlcPaymentRate)
}
\description{
Commodity specific PLC payment rates including statutory and effective reference prices.
}
\details{
To view code used to generate this data set, see \code{./data-raw/fsaArcPlc/supplementary_files/fsaPlcPaymentRates.R}
}
\keyword{datasets}
