% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genFSA.R
\name{genFSA}
\alias{genFSA}
\title{Feasible Solution Algorithm (FSA) for Genetic Models}
\usage{
genFSA(yname, data, fixvar = NULL, quad = F, m = 2, numrs = 1,
  save_solutions = F, cores = 1, interactions = T, criterion = AIC,
  minmax = "min", fam = "binomial", checknum = NA, ...)
}
\arguments{
\item{yname}{the character form of the name of the y (response) variable. Please write the name of the y variable in quotes (example: "Disease").}

\item{data}{a data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{quad}{to include quadratic terms or not.}

\item{numrs}{number of random starts to perform.}

\item{save_solutions}{whether to save the solutions in the current working directory as 'FSAsolutions.csv'.}

\item{cores}{number of cores to use while running. Note: Windows can only use 1 core. See mclapply for details.}

\item{interactions}{T or F for whether to include interactions in model. Defaults to FALSE.}

\item{criterion}{which criterion function to either maximize or minimize. For linear models one can use: apress (Allens Press Statistic), int.p.val (Interaction p-value), AIC, BIC.}

\item{minmax}{whether to minimize or maximize the criterion function}

\item{fam}{family argument passed to glm. A description of the error distribution and link function to be used in the model. This can be a character string naming a family function, a family function or the result of a call to a family function.}

\item{checknum}{Number of models to fit at each swap.}

\item{...}{arguments to be passed to the glm function}

\item{fixvar=NULL}{a variable to fix in the model. Usually a covariate that should always be included (Example: Age, Sex). Will still consider it with interactions.}

\item{m=2}{order of terms to potentially include. If interactions is set to TRUE then m is the order of interactions to be considered. Defaults to 2. For Subset selection (interaction=F), m is the size of the subset to examine.}
}
\value{
returns a list of solutions and table of unique solutions.
$solutions is a matrix of fixed terms, start position, feasible solution, criterion function value (p-value of interaction), and number of swaps to solution.
$table is a matrix of the unique feasible solutions and how many times they occurred out of the number of random starts chosen. It also returns any warning messages with these solutions in the last column.
$efficiency is text comparing how many models you ran during your FSA search compared to how many you would have done with exhaustive search. Note: The FSA algorithm takes additional time to run on top of the model checks that were done during the algorithm. This additional time is approximately 15% more time than if you had just ran the model checks.
}
\description{
A function using a Feasible Solution Algorithm to find a set of feasible solutions for a genetic model of a specific form that could include mth-order interactions (Note that these solutions are optimal in the sense that no one swap to any of the variables will increase the criterion function.)
}
\details{
PLEASE NOTE: make sure categorical variables are factors or characters otherwise answers will not reflect the variable being treated as a continuous variable.
}

