% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fsa_payments.R
\name{get_fsa_payments}
\alias{get_fsa_payments}
\title{Loads data from FSA individual payment files for specified years and programs}
\usage{
get_fsa_payments(
  year = NULL,
  program = NULL,
  year_type = "program",
  aggregation = "national"
)
}
\arguments{
\item{year}{numeric or character values specifying the years of interest (ex: \code{c(2020:2024)})}

\item{program}{a character vector indicating the programs to load. Defaults to all programs.}

\item{year_type}{one of "program", "payment", or "fiscal" which determines the year type to load. Defaults to "program".}

\item{aggregation}{One of "individual", "county", "state", or "national" which determines the level of aggregation.
The level of aggregation also determines the columns returned. For example, the state name would not be returned
if national level data was indicated.}
}
\value{
a data frame
}
\description{
Loads data from FSA individual payment files for specified years and programs
}
\examples{
\dontrun{get_fsa_payments(year = 2024,
                                    program = c("ARC-CO","PLC"),
                                    year_type = "program",
                                    aggregation = "county")}
}
