% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmFSA.R
\name{glmFSA}
\alias{glmFSA}
\title{FSA Generalized Linear Models}
\usage{
glmFSA(formula, data, fixvar = NULL, quad = F, m = 2, numrs = 1,
  save_solutions = T, fam, cores = detectCores()/3, ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under "Details".}

\item{data}{a data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.}

\item{quad}{to include quadratic terms or not.}

\item{numrs}{number of random starts to perform.}

\item{fam}{family argument passed to glm. a description of the error distribution and link function to be used in the model. This can be a character string naming a family function, a family function or the result of a call to a family function.}

\item{cores}{number of cores to use while running. Note: Windows can only use 1 core. See mclapply for details.}

\item{...}{arguments to be passed to the glm function}

\item{fixvar=NULL}{a variable to fix in the model. Usually a covariate that should always be included (Age, Sex, ...). Will still consider it with interactions.}

\item{m=2}{order of interaction to look at. Defults to 2.}

\item{save_results}{whether to save the results in the current working directory as 'FSAresults.csv'.}
}
\value{
a list of results and a table of unique results.
}
\description{
A function using a Feasible Soultion Algorithm to find a set of feasible (optimal in that no one swap to any of the variables will increase the criterion function) mth order interaction Generalized Linear Model
}
\details{
PLEASE NOTE: make sure categorical variables are factors or characters otherwise answers will not reflect the variable being treated as a continuous variable.
}
\examples{
#use mtcars package see help(mtcars)
data(mtcars)
colnames(mtcars)
df<-mtcars
df[,c(2,8:11)]<-lapply(mtcars[,c(2,8:11)],factor)
glmFSA(am~cyl*disp,data=df,fixvar="hp",quad=F,m=2,numrs=10,save_solutions=F,fam="binomial",cores=1)
fit<-glm(am~hp*qsec,data=df,family="binomial") #this is the most common answer from lmFSA.
summary(fit) #review
}

