% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{int.p.val}
\alias{int.p.val}
\title{An rFSA Criterion Function.}
\usage{
int.p.val(model, name = "Interaction P-Value")
}
\arguments{
\item{model}{lm or glm fit to be passed.}

\item{name}{passed to print.FSA}
}
\description{
rFSA Criterion Function to compute Liklihood Ratio Test Statistics p-value for the largest order interation term.
}
