<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/criterion.R
\name{list.criterion}
\alias{list.criterion}
\title{List all included Criteria function for lmFSA and glmFSA.}
\usage{
list.criterion()
}
\value{
list of functions and whether lmFSA or glmFSA work with those functions.
}
\description{
List all included Criteria function for lmFSA and glmFSA.
}
\examples{
list.criterion()
}

=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{list.criterion}
\alias{list.criterion}
\title{List all included Criteria function for lmFSA and glmFSA.}
\usage{
list.criterion()
}
\value{
list of functions and whether lmFSA or glmFSA work with those functions.
}
\description{
List all included Criteria function for lmFSA and glmFSA.
}
\examples{
list.criterion()
}
>>>>>>> 6e2c25ebe54e059b1e387c860c1d0fa8f6e75136
