% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lmFSA.R
\name{lmFSA}
\alias{lmFSA}
\title{FSA Linear Models}
\usage{
lmFSA(formula, data, fixvar = NULL, quad = F, m = 2, numrs = 1,
  save_solutions = F, cores = 1, interactions = T,
  criterion = r.squared, minmax = "max", ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under "Details".}

\item{data}{a data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.}

\item{numrs}{number of random starts to perform.}

\item{save_solutions}{whether to save the solutions in the current working directory as 'FSAsolutions.csv'.}

\item{cores}{number of cores to use while running. Note: Windows can only use 1 core. See mclapply for details.}

\item{interactions}{T or F for whether to include interactions in model.}

\item{criterion}{which criterion function to either maximize or minimize}

\item{minmax}{whether to minimize or maximize the criterion function}

\item{...}{arguments to be passed to the lm function}

\item{fixvar=NULL}{a variable to fix in the model. Usually a covariate that should always be included (Age, Sex, ...). Will still consider it with interactions.}

\item{m=2}{order of interaction to look at. Defults to 2.}
}
\value{
returns a list of solutions and table of unique solutions.
$solutions is a matrix of fixed terms, start position, feasible solution, criterion function value (p-value of interaction), and number of swaps to solution.
$table is a matrix of the unique feasible solutions and how many times they occured out of the number of random starts chosen. It also returns any warning messages with these solutions in the last column.
}
\description{
A function using a Feasible Soultion Algorithm to find a set of feasible solutions for the linear model including mth-order interactions (Note that these solutions are optimal in the sense that no one swap to any of the variables will increase the criterion function.)
}
\details{
PLEASE NOTE: make sure categorical variables are factors or characters otherwise answers will not reflect the variable being treated as a continuous variable.
}
\examples{
#use mtcars package see help(mtcars)
data(mtcars)
colnames(mtcars)
lmFSA(mpg~cyl*disp,data=mtcars,fixvar="hp",quad=F,m=2,numrs=10,save_solutions=F,cores=1)
fit<-lm(yname="mpg",data=mtcars) #this is the most common answer from lmFSA.
summary(fit) #review
}

