% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{max_abs_resid}
\alias{max_abs_resid}
\title{Return maximum absolute residual from a model}
\usage{
max_abs_resid(model)
}
\arguments{
\item{model}{model obj}
}
\description{
Return maximum absolute residual from a model
}
