% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{nextswap}
\alias{nextswap}
\title{Variables to include in the >1st step of an mth order interaction model determined from the Feasible Soution Alorithm.}
\usage{
nextswap(curpos, n, prevpos, quad)
}
\arguments{
\item{curpos}{A vector of length greater than 2 of what current explantory varialbes are being used in the model.}

\item{n}{The number of explanatory variables in available to swap.}

\item{prevpos}{A vector of previous best spots}

\item{quad}{Whether to include quadratic terms. ie (x1*x1) as potential swaps.}
}
\value{
a matrix with the possible forms by column.
}
\description{
Finds the swaps available given a current position given previous picks.
}
