% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.FSA.R
\name{print.FSA}
\alias{print.FSA}
\title{Printing function for FSA solutions}
\usage{
\method{print}{FSA}(x, ...)
}
\arguments{
\item{x}{FSA object to print details about.}

\item{...}{arguments to be passed to other functions.}
}
\value{
list of Feasible Solution Formula's, Original Fitted model formula and criterion function and times converged to details.
}
\description{
Printing function for FSA solutions
}
\examples{
#use mtcars package see help(mtcars)
data(mtcars)
colnames(mtcars)
fit<-lmFSA(formula="mpg~hp*wt",data=mtcars,fixvar="hp",
                   quad=FALSE,m=2,numrs=10,save_solutions=FALSE,cores=1)
print(fit)
}
