% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{QICu.geeglm}
\alias{QICu.geeglm}
\title{Return QICu for geepack::geeglm}
\usage{
QICu.geeglm(gee.obj)
}
\arguments{
\item{gee.obj}{geeglm obj}
}
\description{
Computes quasi-likelihood under the independence criterion (QICu)
}
