% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{r.squared}
\alias{r.squared}
\title{An rFSA Criterion Function.}
\usage{
r.squared(model, name = "R Squared")
}
\arguments{
\item{model}{lm or glm fit to be passed.}

\item{name}{passed to print.FSA}
}
\description{
rFSA Criterion Function to compute R squared.
}
