% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{rmse}
\alias{rmse}
\title{An rFSA Criterion Function.}
\usage{
rmse(model, name = "RMSE")
}
\arguments{
\item{model}{lm or glm fit to be passed.}

\item{name}{passed to print.FSA}
}
\description{
rFSA Criterion Function to compute Root Mean Squared Error.
}
