% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstart.R
\name{rstart}
\alias{rstart}
\title{Random starts with m variables}
\usage{
rstart(m, nvars, quad = FALSE, numrs = 1)
}
\arguments{
\item{m}{order of interaction to include.}

\item{nvars}{number of varialbes that could be included in the interaction.}

\item{quad}{whether to include quadratic terms or not.}

\item{numrs}{number of random starts to compute.}
}
\value{
A matrix of random starts by row
}
\description{
Random starts with m variables
}
\examples{
rstart(2,5,quad=FALSE,numrs=3) #doesn't include quadratic terms
rstart(2,5,quad=TRUE,numrs=3) #does include quadratic terms
}

