% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.FSA.R
\name{summary.FSA}
\alias{summary.FSA}
\title{Summary function for FSA solutions}
\usage{
\method{summary}{FSA}(object, ...)
}
\arguments{
\item{object}{FSA object to see summaries on.}

\item{...}{arguments to be passed to other functions.}
}
\value{
list of summarized lm or glm output.
}
\description{
Summary function for FSA solutions
}
\examples{
#use mtcars package see help(mtcars)
data(mtcars)
colnames(mtcars)
fit<-lmFSA(formula="mpg~hp*wt",data=mtcars,fixvar="hp",
           quad=FALSE,m=2,numrs=10,save_solutions=FALSE,cores=1)
summary(fit)
}
