% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twFSA.R
\name{twFSA}
\alias{twFSA}
\title{twFSA}
\usage{
twFSA(formula, data, fitfunc = lm, fixvar = NULL, quad = FALSE,
  cores = 1, criterion = AIC, minmax = "min", checkfeas = NULL,
  var4int = NULL, min.nonmissing = 1, ...)
}
\description{
twFSA
}
