% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{bdist}
\alias{bdist}
\title{An rFSA Criterion Function.}
\usage{
bdist(model, name = "B Distance")
}
\arguments{
\item{model}{lm or glm fit to be passed.}

\item{name}{passed to print.FSA}
}
\description{
rFSA Criterion Function to compute the Bhattacharyya distance.
}
