% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clear_rfsa_cache}
\alias{clear_rfsa_cache}
\title{Clear the package cache of downloaded RDS files}
\usage{
clear_rfsa_cache()
}
\value{
Invisibly returns \code{NULL}. A message is printed indicating which
directory was cleared.
}
\description{
Deletes the entire cache directory used by the \strong{rfsa} package to store
downloaded \*.rds files. Useful if you need to force re-download of data,
or free up disk space.
}
\examples{
\dontrun{
# Remove all cached RDS files so they will be re-downloaded on next use
clear_rfsa_cache()
}
}
