% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fsaArcCoPrice}
\alias{fsaArcCoPrice}
\title{ARC-CO benchmark prices.}
\format{
A tibble
\describe{
\item{Commodity}{Crop name or commodity covered by ARC-CO and MYA prices.}
\item{Marketing Year}{The dates that define the commodity specific marketing year (e.x. Sep. 1-Aug. 31).}
\item{Publishing Dates for the Final T-0 MYA Price and T-0 Actual ARC-CO Price}{Date the final prices for the marketing year were published.}
\item{Unit}{Unit of measure for the commodity (e.g., Bushel, Pound).}
\item{Statutory Reference Price}{Reference price set in legislation, used in ARC and PLC payment calculations.}
\item{Final T-5 Annual Benchmark Price}{Price from five years prior to the marketing year, used in ARC-CO benchmark calculation.}
\item{Final T-4 Annual Benchmark Price}{Price from four years prior to the marketing year, used in ARC-CO benchmark calculation.}
\item{Final T-3 Annual Benchmark Price}{Price from three years prior to the marketing year, used in ARC-CO benchmark calculation.}
\item{Final T-2 Annual Benchmark Price}{Price from two years prior to the marketing year, used in ARC-CO benchmark calculation.}
\item{Final T-1 Annual Benchmark Price}{Price from one year prior to the marketing year, used in ARC-CO benchmark calculation.}
\item{Final T-0 ARC-CO Benchmark Price}{ARC-CO benchmark price for the current marketing year.}
\item{Final T-0 MYA Price}{Marketing Year Average price for the current year.}
\item{T-0 National Loan Rate}{National loan rate for the current marketing year.}
\item{Final T-0 Actual ARC-CO Price}{Actual price used in ARC-CO calculations}
\item{year}{Marketing year (e.g., "2020-2021").}
}
}
\source{
\url{https://www.fsa.usda.gov/resources/programs/arc-plc/program-data}
}
\usage{
data(fsaArcCoPrice)
}
\description{
A dataset containing annual commodity-level price data used in the calculation of ARC-CO (Agriculture Risk Coverage - County) benchmark prices.
}
\keyword{datasets}
