% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fsaArcPlcBaseAcres}
\alias{fsaArcPlcBaseAcres}
\title{ARC/PLC enrolled base acres by commodity}
\format{
A data frame with 225 rows and 14 variables:
\describe{
\item{covered_commodity}{Name of the covered commodity}
\item{plc_covered_commodity_contract_base}{Acres enrolled in PLC with covered commodity base}
\item{plc_plantings_attributed_to_generic_base}{Acres assigned to PLC from generic base}
\item{arc_co_covered_commodity_contract_base}{Acres enrolled in ARC-CO with covered commodity base}
\item{arc_co_plantings_attributed_to_generic_base}{Acres assigned to ARC-CO from generic base}
\item{arc_ic_enrolled_base_covered_commodity_contract_base}{Acres enrolled in ARC-IC with covered commodity base}
\item{total}{Total base acres enrolled for the commodity in all programs}
\item{plc_total}{Total PLC base acres}
\item{arc_co_total}{Total ARC-CO base acres}
\item{arc_ic_total}{Total ARC-IC base acres}
\item{year}{Year of enrollment}
\item{arc_co_all}{ARC-CO acreage (all base types), available for 2018 onward}
\item{arc_co_irrigated}{Irrigated ARC-CO acreage, available for 2018 onward}
\item{arc_co_nonirrigated}{Nonirrigated ARC-CO acreage, available for 2018 onward}
}
}
\source{
\url{https://www.fsa.usda.gov/resources/programs/arc-plc/program-data}
}
\usage{
data(fsaArcPlcBaseAcres)
}
\description{
This dataset provides annual  base acreage by program and commodity under the ARC-CO, ARC-IC, and PLC programs
administered by the USDA Farm Service Agency. It includes acreage originally attributed to generic base, as well as total
enrolled acreage by commodity and year. From 2018 onward, base assignments under the Seed Cotton program are also reflected.
}
\keyword{datasets}
