% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmFSA.R
\name{glmFSA}
\alias{glmFSA}
\title{FSA Generalized Linear Models}
\usage{
glmFSA(formula, data, fixvar = NULL, quad = F, m = 2, numrs = 1,
  save_solutions = T, fam, cores = 1, interactions = T, ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under "Details".}

\item{data}{a data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.}

\item{quad}{to include quadratic terms or not.}

\item{numrs}{number of random starts to perform.}

\item{fam}{family argument passed to glm. a description of the error distribution and link function to be used in the model. This can be a character string naming a family function, a family function or the result of a call to a family function.}

\item{cores}{number of cores to use while running. Note: Windows can only use 1 core. See mclapply for details.}

\item{interactions}{T or F for whether to include interactions in model.}

\item{...}{arguments to be passed to the glm function}

\item{fixvar=NULL}{a variable to fix in the model. Usually a covariate that should always be included (Age, Sex, ...). Will still consider it with interactions.}

\item{m=2}{order of interaction to look at. Defults to 2.}

\item{save_results}{whether to save the results in the current working directory as 'FSAresults.csv'.}
}
\value{
returns a list of solutions and table of unique solutions.
$solutions is a matrix of fixed terms, start position, feasible solution, criterion function value (p-value of interaction), and number of swaps to solution.
$table is a matrix of the unique feasible solutions and how many times they occured out of the number of random starts chosen. It also returns any warning messages with these solutions in the last column.
}
\description{
A function using a Feasible Soultion Algorithm to find a set of feasible (optimal in that no one swap to any of the variables will increase the criterion function) mth order interaction Generalized Linear Model
}
\details{
PLEASE NOTE: make sure categorical variables are factors or characters otherwise answers will not reflect the variable being treated as a continuous variable.
}
\examples{
dat<-read.csv("https://archive.ics.uci.edu/ml/machine-learning-databases/hepatitis/hepatitis.data",header = F)
colnames(dat)<-c("Class","Age","Sex","Sterioid","Antivirals","Fatigue","Malaise","Anorexia","Liver Big","Liver Firm","Spleen Palpable","Spiders","Ascites","Varices","Bilirubin","Alk Phosphate","Sgot","Albumin","Protime","Histology")
dat<-as.matrix(dat)
dat[which(dat=="?")]=NA
dat<-data.frame(dat)
dat[,c(2,15,16,17,18,19)]<-lapply(X = dat[,c(2,15,16,17,18,19)],as.numeric)
colnames(dat)
glmFSA(Class~Age*Sex,data=dat,fixvar="Age",quad=F,m=2,numrs=10,save_solutions = F,fam="binomial",cores=1)

}

