% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{list.criterion}
\alias{list.criterion}
\title{List all included Criteria function for lmFSA and glmFSA.}
\usage{
list.criterion()
}
\value{
list of functions and whether lmFSA or glmFSA work with those functions.
}
\description{
List all included Criteria function for lmFSA and glmFSA.
}
\examples{
list.criterion()
}
