% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{which.max.na}
\alias{which.max.na}
\title{An rFSA Internal Function.}
\usage{
which.max.na(vec)
}
\arguments{
\item{vec}{Vector to be passed.}
}
\description{
rFSA function to compute the maximum value from a vector with NA's.
}
