% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{which.min.na}
\alias{which.min.na}
\title{An rFSA Internal Function.}
\usage{
which.min.na(vec)
}
\arguments{
\item{vec}{Vector to be passed.}
}
\description{
rFSA function to compute the minimum value from a vector with NA's.
}
